Inclusão de classe para tratamento de erros
