+++ src/main/java/med/voll/apirestclinicamedica/controller/PacienteController.java
@@ -1,7 +1,7 @@
 package med.voll.apirestclinicamedica.controller;
 
 import jakarta.validation.Valid;
-import med.voll.apirestclinicamedica.paciente.*;
+import med.voll.apirestclinicamedica.domain.paciente.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
