+++ src/main/java/med/voll/apirestclinicamedica/domain/medico/DadosCadastroMedico.java
@@ -1,13 +1,13 @@
-package med.voll.apirestclinicamedica.paciente;
+package med.voll.apirestclinicamedica.domain.medico;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
-import med.voll.apirestclinicamedica.endereco.DadosEndereco;
+import med.voll.apirestclinicamedica.domain.endereco.DadosEndereco;
 
-public record DadosCadastroPaciente(
+public record DadosCadastroMedico(
         @NotBlank
         String nome,
         @NotBlank
@@ -16,8 +16,10 @@ public record DadosCadastroPaciente(
         @NotBlank
         String telefone,
         @NotBlank
-        @Pattern(regexp = "\\d{3}\\.\\d{3}\\.\\d{3}\\-\\d{2}")
-        String cpf,
+        @Pattern(regexp = "\\d{4,6}")
+        String crm,
+        @NotNull
+        Especialidade especialidade,
         @NotNull
         @Valid
         DadosEndereco endereco
