+++ src/main/java/med/voll/apirestclinicamedica/domain/paciente/DadosCadastroPaciente.java
@@ -1,13 +1,13 @@
-package med.voll.apirestclinicamedica.medico;
+package med.voll.apirestclinicamedica.domain.paciente;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
-import med.voll.apirestclinicamedica.endereco.DadosEndereco;
+import med.voll.apirestclinicamedica.domain.endereco.DadosEndereco;
 
-public record DadosCadastroMedico(
+public record DadosCadastroPaciente(
         @NotBlank
         String nome,
         @NotBlank
@@ -16,10 +16,8 @@ public record DadosCadastroMedico(
         @NotBlank
         String telefone,
         @NotBlank
-        @Pattern(regexp = "\\d{4,6}")
-        String crm,
-        @NotNull
-        Especialidade especialidade,
+        @Pattern(regexp = "\\d{3}\\.\\d{3}\\.\\d{3}\\-\\d{2}")
+        String cpf,
         @NotNull
         @Valid
         DadosEndereco endereco
