+++ src/main/java/med/voll/apirestclinicamedica/infra/TratadorDeErros.java
@@ -1,0 +1,36 @@
+package med.voll.apirestclinicamedica.infra;
+
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.sql.SQLIntegrityConstraintViolationException;
+
+@RestControllerAdvice
+public class TratadorDeErros {
+
+    @ExceptionHandler(EntityNotFoundException.class)
+    public ResponseEntity tratarErro404() {
+        return ResponseEntity.notFound().build();
+    }
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity tratarErro400(MethodArgumentNotValidException ex) {
+        var erros = ex.getFieldErrors();
+        return ResponseEntity.badRequest().body(erros.stream().map(DadosErroValidacao::new).toList());
+    }
+
+    private record DadosErroValidacao(String nome, String mensagem) {
+        public DadosErroValidacao(FieldError erro) {
+            this(
+                    erro.getField(),
+                    erro.getDefaultMessage()
+            );
+        }
+    }
+}
